third_party/WebKit/LayoutTests/editing/spelling/spelling-changed-text.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script src="spellcheck_test.js"></script>
<script>
spellcheck_test(
    '<div contenteditable>|</div>',
    'insertText Spell wellcome. Is it broken?',
    '<div contenteditable>Spell #wellcome#. Is it broken?</div>',
    {
      title: 'Setup a content editable div with spelling marker',
      callback: sample => spellcheck_test(
          sample,
          document => {
            assert_not_equals(window.eventSender, undefined, 'This test requires eventSender.');

            const destination = document.querySelector('div');
            const textNode = destination.firstChild;

            // Select the text "Is it broken?".
            const deleteRange = document.createRange();
            deleteRange.setStart(textNode, 15);
            deleteRange.setEnd(textNode, 29);
            document.getSelection().removeAllRanges();
            document.getSelection().addRange(deleteRange);

            // Del key to delete the text "Is it broken?".
            eventSender.keyDown("Delete", null);

            // Context click to show the context menu.
            var x = destination.offsetParent.offsetLeft + destination.offsetLeft + 50;
            var y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
            eventSender.mouseMoveTo(x, y);
            contextMenuElements = eventSender.contextClick();

            // Esc key to hide the context menu.
            eventSender.keyDown("Escape", null);
          },
          '<div contenteditable>Spell #wellcome#.</div>',
          'Spellcheck should not crash after the text has changed.')
    });
</script>
</body>
</html>
